Add files via upload
We can use Spark's APIs to read both structured and unstructured data from MySQL and MongoDB databases, transform the data as needed, and write it to HDFS in the desired format. Spark supports various file formats for writing data to HDFS, such as CSV, Parquet, and Avro.```

By using Spark, you can process large volumes of data efficiently and easily perform complex transformations on the data before writing it to HDFS. This can be useful for a variety of big data applications, such as data processing pipelines, machine learning, and data analysis.